+++ examples/openapi-ts-axios/src/App.tsx
@@ -1,6 +1,5 @@
 import './App.css';
 
-import { createClient } from '@hey-api/client-axios';
 import * as Form from '@radix-ui/react-form';
 import { DownloadIcon, PlusIcon, ReloadIcon } from '@radix-ui/react-icons';
 import {
@@ -15,6 +14,7 @@ import {
   Text,
   TextField,
 } from '@radix-ui/themes';
+import { createClient } from '@ts-sdk-gen/client-axios';
 import { useState } from 'react';
 
 import { PetSchema } from './client/schemas.gen';