+++ packages/openapi-ts/src/plugins/@tanstack/query-core/plugin-legacy/constants.ts
@@ -60,7 +60,9 @@ export const toQueryKeyName = ({
 
 export const getClientBaseUrlKey = () => {
   const config = getConfig();
-  return config.client.name === '@hey-api/client-axios' ? 'baseURL' : 'baseUrl';
+  return config.client.name === '@ts-sdk-gen/client-axios'
+    ? 'baseURL'
+    : 'baseUrl';
 };
 
 export const getPaginationIn = (parameter: OperationParameter) => {