+++ packages/openapi-ts/src/types/config.ts
@@ -7,7 +7,7 @@ import type {
 } from './utils';
 
 export const CLIENTS = [
-  '@hey-api/client-axios',
+  '@ts-sdk-gen/client-axios',
   '@ts-sdk-gen/client-axios',
   '@ts-sdk-gen/client-fetch',
   '@ts-sdk-gen/client-fetch',