+++ examples/openapi-ts-fastify/CHANGELOG.md
@@ -5,11 +5,11 @@
 ### Patch Changes
 
 - Updated dependencies [[`734a62d`](https://github.com/ts-sdk-gen/openapi-ts/commit/734a62dd8d594b8266964fe16766a481d37eb7df), [`734a62d`](https://github.com/ts-sdk-gen/openapi-ts/commit/734a62dd8d594b8266964fe16766a481d37eb7df)]:
-  - @ts-sdk-gengen/client-fetch@0.5.0
+  - @ts-sdk-gen/client-fetch@0.5.0
 
 ## 0.0.2
 
 ### Patch Changes
 
 - Updated dependencies [[`4c853d0`](https://github.com/ts-sdk-gen/openapi-ts/commit/4c853d090b79245854d13831f64731db4a92978b)]:
-  - @ts-sdk-gengen/client-fetch@0.4.4
+  - @ts-sdk-gen/client-fetch@0.4.4